Moved Projectile into modelPieces
It's just more logical to group it there, especially now that CubicCurve is there too.
