+++ src/main/java/drgtools/dpscalc/drivers/GuiController.java
@@ -65,7 +65,6 @@ import drgtools.dpscalc.weapons.scout.zhukov.Zhukov;
 	TODO List
 	1. Move calculateNumEnemies, calculateDamagePerHit, and some Utility stuff into DamageComponent
 	3. Add some prettyPrint/toString methods to the complex objects, to facilitate easier debugging in the future.
-	5. Expand upon the CurveFloat object, now that I have the formula to extract the polynomial equations.
 	6. PR & merge damage-refactor
 	7. Figure out how to do Subata first with the newly refactored mechanics, and then use that paradigm to update the other models.
 	8. Implement STEs that can stack with themself multiple times (Neuro-Lasso's slow, Cryo Bolt's cold/sec, etc)
