+++ src/main/java/drgtools/dpscalc/modelPieces/Projectile.java
@@ -1,4 +1,4 @@
-package drgtools.dpscalc.weapons;
+package drgtools.dpscalc.modelPieces;
 
 import drgtools.dpscalc.modelPieces.damage.DamageInstance;
