+++ src/main/java/drgtools/dpscalc/weapons/driller/epc/PRJ_ChargedShot.java
@@ -1,7 +1,7 @@
 package drgtools.dpscalc.weapons.driller.epc;
 
 import drgtools.dpscalc.modelPieces.damage.DamageComponent;
-import drgtools.dpscalc.weapons.Projectile;
+import drgtools.dpscalc.modelPieces.Projectile;
 
 public class PRJ_ChargedShot extends Projectile {
     public PRJ_ChargedShot(DamageComponent dmg) {
