+++ src/main/java/drgtools/dpscalc/weapons/driller/sludgePump/PRJ_RegularShot.java
@@ -1,7 +1,7 @@
 package drgtools.dpscalc.weapons.driller.sludgePump;
 
 import drgtools.dpscalc.modelPieces.damage.DamageComponent;
-import drgtools.dpscalc.weapons.Projectile;
+import drgtools.dpscalc.modelPieces.Projectile;
 
 public class PRJ_RegularShot extends Projectile {
     public PRJ_RegularShot(double velocity, double gravity, DamageComponent dmg) {
